Make fifo lifo a gem
