+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+## [Unreleased]
+
+## [0.1.0] - 2022-02-05
+
+- Initial release
