+++ lib/fifo_lifo/fifo.rb
