+++ lib/fifo_lifo/lifo.rb
