+++ lib/fifo_lifo/utils.rb
