+++ spec/fifo_lifo_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+RSpec.describe FifoLifo do
+  it "has a version number" do
+    expect(FifoLifo::VERSION).not_to be nil
+  end
+
+  it "does something useful" do
+    expect(false).to eq(true)
+  end
+end
